move non-essential files out of src/
