--- src/validators.js
@@ -1,288 +1,0 @@
-import validate, { walkValidationTree } from './validators/validate'
-import DateControlValidator from './validators/datecontrol'
-import DateRangeValidator from './validators/daterange'
-import BankruptcyValidator, { BankruptcyItemValidator } from './validators/bankruptcy'
-import BirthPlaceValidator from './validators/birthplace'
-import IdentificationContactInformationValidator, { ContactEmailValidator, ContactPhoneNumberValidator } from './validators/identificationcontacts'
-import IdentificationValidator from './validators/identification'
-import IdentificationNameValidator from './validators/identificationname'
-import IdentificationBirthDateValidator from './validators/identificationbirthdate'
-import IdentificationBirthPlaceValidator from './validators/identificationbirthplace'
-import IdentificationSSNValidator from './validators/identificationssn'
-import NameValidator from './validators/name'
-import IdentificationOtherNamesValidator, { OtherNameValidator } from './validators/identificationothernames'
-import PassportValidator from './validators/passport'
-import IdentificationPhysicalValidator from './validators/identificationphysical'
-import GamblingValidator, { GamblingItemValidator } from './validators/gambling'
-import HistoryResidenceValidator, { ResidenceValidator } from './validators/residence'
-import HistoryEmploymentValidator, { EmploymentValidator } from './validators/employment'
-import HistoryEducationValidator, { EducationItemValidator } from './validators/education'
-import FederalServiceValidator, { FederalServiceItemValidator } from './validators/federalservice'
-import SelectiveServiceValidator, { hideSelectiveService } from './validators/selectiveservice'
-import MilitaryHistoryValidator, { MilitaryServiceValidator} from './validators/militaryhistory'
-import MilitaryDisciplinaryValidator, { ProcedureValidator, hideDisciplinaryProcedures } from './validators/militarydisciplinary'
-import MilitaryForeignValidator, { ForeignServiceValidator, ForeignContactValidator } from './validators/militaryforeign'
-import PoliceOffensesValidator from './validators/policeoffenses'
-import PoliceOtherOffensesValidator from './validators/policeotheroffenses'
-import OtherOffenseValidator from './validators/otheroffense'
-import DomesticViolenceValidator, { DomesticViolenceItem } from './validators/domesticviolence'
-import OffenseValidator from './validators/offense'
-import RelativesValidator, { RelativeValidator, AliasValidator } from './validators/relatives'
-import ConsultationValidator from './validators/consultation'
-import TaxesValidator, { TaxValidator } from './validators/taxes'
-import CardAbuseValidator, { CardAbuseItemValidator } from './validators/cardabuse'
-import CreditValidator, { CreditItemValidator } from './validators/credit'
-import HospitalizationsValidator, { HospitalizationValidator } from './validators/hospitalization'
-import NonpaymentValidator, { NonpaymentItemValidator } from './validators/nonpayment'
-import DelinquentValidator, { DelinquentItemValidator } from './validators/delinquent'
-import DiagnosesValidator from './validators/diagnoses'
-import DiagnosisValidator, { ExistingConditionsDiagnosisValidator } from './validators/diagnosis'
-import ExistingConditionsValidator from './validators/existingconditions'
-import PsychologicalValidator, { hideExistingConditions } from './validators/psychological'
-import CompetenceValidator from './validators/competence'
-import TreatmentValidator from './validators/treatment'
-import CitizenshipValidator from './validators/citizenship'
-import CitizenshipMultipleValidator, { CitizenshipItemValidator } from './validators/citizenship-multiple'
-import CitizenshipPassportsValidator, { TravelItemValidator } from './validators/citizenship-passports'
-import CivilUnionValidator from './validators/civilunion'
-import MaritalValidator from './validators/marital'
-import CohabitantsValidator, { CohabitantValidator } from './validators/cohabitant'
-import RelationshipsValidator from './validators/relationships'
-import DivorceValidator from './validators/divorce'
-import PeopleValidator from './validators/people'
-import PersonValidator from './validators/person'
-import ForeignContactsValidator, { ForeignNationalValidator } from './validators/foreigncontacts'
-import ForeignDirectActivityValidator from './validators/foreigndirectactivity'
-import ForeignDirectInterestValidator from './validators/foreigndirectinterest'
-import ForeignIndirectActivityValidator from './validators/foreignindirectactivity'
-import ForeignIndirectInterestValidator from './validators/foreignindirectinterest'
-import ForeignRealEstateActivityValidator from './validators/foreignrealestateactivity'
-import ForeignRealEstateInterestValidator from './validators/foreignrealestateinterest'
-import ForeignBenefitActivityValidator from './validators/foreignbenefitactivity'
-import LegalNonCriminalCourtActionsValidator, { NonCriminalCourtActionValidator } from './validators/legalnoncriminalcourtactions'
-import LegalInvestigationsHistoryValidator, { HistoryValidator } from './validators/legalinvestigationshistory'
-import LegalInvestigationsRevokedValidator, { RevokedValidator } from './validators/legalinvestigationsrevoked'
-import LegalInvestigationsDebarredValidator, { DebarredValidator } from './validators/legalinvestigationsdebarred'
-import LegalTechnologyUnauthorizedValidator, { UnauthorizedValidator } from './validators/legaltechnologyunauthorized'
-import LegalTechnologyManipulatingValidator, { ManipulatingValidator } from './validators/legaltechnologymanipulating'
-import LegalTechnologyUnlawfulValidator, { UnlawfulValidator } from './validators/legaltechnologyunlawful'
-import LegalAssociationsTerroristValidator, { TerroristValidator } from './validators/legalassociationsterrorist'
-import ForeignBenefitValidator from './validators/foreignbenefit'
-import ForeignActivitiesSupportValidator, { SupportValidator } from './validators/foreignsupport'
-import ForeignBornDocumentValidator from './validators/foreignborndocument'
-import ForeignBusinessAdviceValidator, { AdviceValidator } from './validators/foreignbusinessadvice'
-import ForeignBusinessFamilyValidator, { FamilyValidator } from './validators/foreignbusinessfamily'
-import ForeignBusinessEmploymentValidator, { ForeignBusinessEmploymentItemValidator } from './validators/foreignbusinessemployment'
-import ForeignBusinessVenturesValidator, { VenturesValidator } from './validators/foreignbusinessventures'
-import ForeignBusinessConferencesValidator, { ConferencesValidator } from './validators/foreignbusinessconferences'
-import ForeignBusinessContactValidator, { ContactValidator } from './validators/foreignbusinesscontact'
-import ForeignBusinessSponsorshipValidator, { SponsorshipValidator } from './validators/foreignbusinesssponsorship'
-import ForeignBusinessPoliticalValidator, { PoliticalValidator } from './validators/foreignbusinesspolitical'
-import ForeignBusinessVotingValidator, { VotingValidator } from './validators/foreignbusinessvoting'
-import ForeignTravelValidator, { TravelValidator } from './validators/foreigntravel.js'
-import AlcoholNegativeImpactsValidator, { NegativeImpactValidator } from './validators/alcoholnegativeimpact'
-import AlcoholOrderedCounselingsValidator, { OrderedCounselingValidator } from './validators/alcoholorderedcounseling'
-import AlcoholVoluntaryCounselingsValidator, { VoluntaryCounselingValidator } from './validators/alcoholvoluntarycounseling'
-import AlcoholReceivedCounselingsValidator, { ReceivedCounselingValidator } from './validators/alcoholreceivedcounseling'
-import DrugUsesValidator, { DrugUseValidator } from './validators/druguses'
-import DrugInvolvementsValidator, { DrugInvolvementValidator } from './validators/druginvolvements'
-import DrugClearanceUsesValidator, { DrugClearanceUseValidator } from './validators/drugclearanceuses'
-import DrugPublicSafetyUsesValidator, { DrugPublicSafetyUseValidator } from './validators/drugpublicsafetyuses'
-import DrugPrescriptionUsesValidator, { DrugPrescriptionUseValidator } from './validators/drugprescriptionuses'
-import DrugOrderedTreatmentsValidator, { DrugOrderedTreatmentValidator } from './validators/drugorderedtreatments'
-import DrugVoluntaryTreatmentsValidator, { DrugVoluntaryTreatmentValidator } from './validators/drugvoluntarytreatments'
-import LegalAssociationsOverthrowValidator, { OverthrowValidator } from './validators/legalassociationsoverthrow'
-import LegalAssociationsViolenceValidator, { ViolenceValidator } from './validators/legalassociationsviolence'
-import LegalAssociationsEngagedValidator, { EngagedValidator } from './validators/legalassociationsengaged'
-import LegalAssociationsAdvocatingValidator, { AdvocatingValidator } from './validators/legalassociationsadvocating'
-import LegalAssociationsActivitiesValidator, { ActivitiesValidator } from './validators/legalassociationsactivities'
-import LegalAssociationsTerrorismValidator from './validators/legalassociationsterrorism'
-import LocationValidator from './validators/location'
-import SignatureValidator from './validators/signature'
-import SentenceValidator from './validators/sentence'
-import { nameIsEmpty, validBranch, validGenericTextfield, validPhoneNumber, validSSN } from './validators/helpers'
-import OrderValidator, { CompetenceOrderValidator, ConsultationOrderValidator } from './validators/order'
-import { hideReleases, hideHippa, formIsSigned, formIsLocked } from './validators/releases'
-
-export default validate
-export {
-  walkValidationTree,
-  DateControlValidator,
-  DateRangeValidator,
-  IdentificationNameValidator,
-  IdentificationBirthDateValidator,
-  IdentificationBirthPlaceValidator,
-  IdentificationSSNValidator,
-  BankruptcyValidator,
-  BankruptcyItemValidator,
-  BirthPlaceValidator,
-  IdentificationContactInformationValidator,
-  ContactEmailValidator,
-  ContactPhoneNumberValidator,
-  IdentificationValidator,
-  NameValidator,
-  IdentificationOtherNamesValidator,
-  OtherNameValidator,
-  PassportValidator,
-  IdentificationPhysicalValidator,
-  GamblingValidator,
-  GamblingItemValidator,
-  HistoryResidenceValidator,
-  ResidenceValidator,
-  HistoryEmploymentValidator,
-  EmploymentValidator,
-  HistoryEducationValidator,
-  EducationItemValidator,
-  SelectiveServiceValidator,
-  MilitaryHistoryValidator,
-  MilitaryServiceValidator,
-  MilitaryDisciplinaryValidator,
-  ProcedureValidator,
-  MilitaryForeignValidator,
-  ForeignServiceValidator,
-  ForeignContactValidator,
-  PoliceOffensesValidator,
-  PoliceOtherOffensesValidator,
-  OtherOffenseValidator,
-  DomesticViolenceValidator,
-  DomesticViolenceItem,
-  OffenseValidator,
-  FederalServiceValidator,
-  FederalServiceItemValidator,
-  RelativesValidator,
-  RelativeValidator,
-  AliasValidator,
-  ConsultationValidator,
-  TaxesValidator,
-  TaxValidator,
-  CardAbuseValidator,
-  CardAbuseItemValidator,
-  CreditValidator,
-  CreditItemValidator,
-  HospitalizationsValidator,
-  NonpaymentValidator,
-  NonpaymentItemValidator,
-  DelinquentValidator,
-  DelinquentItemValidator,
-  DiagnosesValidator,
-  ExistingConditionsDiagnosisValidator,
-  DiagnosisValidator,
-  ExistingConditionsValidator,
-  PsychologicalValidator,
-  CompetenceValidator,
-  TreatmentValidator,
-  HospitalizationValidator,
-  CitizenshipValidator,
-  CitizenshipMultipleValidator,
-  CitizenshipItemValidator,
-  CitizenshipPassportsValidator,
-  TravelItemValidator,
-  CivilUnionValidator,
-  MaritalValidator,
-  CohabitantsValidator,
-  CohabitantValidator,
-  RelationshipsValidator,
-  DivorceValidator,
-  PeopleValidator,
-  PersonValidator,
-  ForeignBornDocumentValidator,
-  ForeignContactsValidator,
-  ForeignNationalValidator,
-  ForeignDirectActivityValidator,
-  ForeignDirectInterestValidator,
-  ForeignIndirectActivityValidator,
-  ForeignIndirectInterestValidator,
-  ForeignRealEstateActivityValidator,
-  ForeignRealEstateInterestValidator,
-  ForeignBenefitActivityValidator,
-  ForeignBenefitValidator,
-  ForeignActivitiesSupportValidator,
-  SupportValidator,
-  ForeignBusinessAdviceValidator,
-  AdviceValidator,
-  ForeignBusinessFamilyValidator,
-  FamilyValidator,
-  ForeignBusinessEmploymentValidator,
-  ForeignBusinessEmploymentItemValidator,
-  ForeignBusinessVenturesValidator,
-  VenturesValidator,
-  ForeignBusinessConferencesValidator,
-  ConferencesValidator,
-  ForeignBusinessContactValidator,
-  ContactValidator,
-  ForeignBusinessSponsorshipValidator,
-  SponsorshipValidator,
-  ForeignBusinessPoliticalValidator,
-  PoliticalValidator,
-  ForeignBusinessVotingValidator,
-  VotingValidator,
-  ForeignTravelValidator,
-  TravelValidator,
-  AlcoholNegativeImpactsValidator,
-  NegativeImpactValidator,
-  AlcoholOrderedCounselingsValidator,
-  OrderedCounselingValidator,
-  AlcoholVoluntaryCounselingsValidator,
-  VoluntaryCounselingValidator,
-  AlcoholReceivedCounselingsValidator,
-  ReceivedCounselingValidator,
-  DrugUsesValidator,
-  DrugUseValidator,
-  DrugInvolvementsValidator,
-  DrugInvolvementValidator,
-  DrugClearanceUsesValidator,
-  DrugClearanceUseValidator,
-  DrugPublicSafetyUsesValidator,
-  DrugPublicSafetyUseValidator,
-  DrugPrescriptionUsesValidator,
-  DrugPrescriptionUseValidator,
-  DrugOrderedTreatmentsValidator,
-  DrugOrderedTreatmentValidator,
-  DrugVoluntaryTreatmentsValidator,
-  DrugVoluntaryTreatmentValidator,
-  LegalNonCriminalCourtActionsValidator,
-  NonCriminalCourtActionValidator,
-  LegalInvestigationsHistoryValidator,
-  HistoryValidator,
-  LegalInvestigationsRevokedValidator,
-  RevokedValidator,
-  LegalInvestigationsDebarredValidator,
-  DebarredValidator,
-  LegalTechnologyUnauthorizedValidator,
-  UnauthorizedValidator,
-  LegalTechnologyManipulatingValidator,
-  ManipulatingValidator,
-  LegalTechnologyUnlawfulValidator,
-  UnlawfulValidator,
-  LegalAssociationsTerroristValidator,
-  TerroristValidator,
-  LegalAssociationsOverthrowValidator,
-  OverthrowValidator,
-  LegalAssociationsViolenceValidator,
-  ViolenceValidator,
-  LegalAssociationsEngagedValidator,
-  EngagedValidator,
-  LegalAssociationsAdvocatingValidator,
-  AdvocatingValidator,
-  LegalAssociationsActivitiesValidator,
-  ActivitiesValidator,
-  LegalAssociationsTerrorismValidator,
-  LocationValidator,
-  SignatureValidator,
-  SentenceValidator,
-  nameIsEmpty,
-  validBranch,
-  validGenericTextfield,
-  validPhoneNumber,
-  validSSN,
-  OrderValidator,
-  CompetenceOrderValidator,
-  ConsultationOrderValidator,
-  hideDisciplinaryProcedures,
-  hideExistingConditions,
-  hideSelectiveService,
-  hideReleases,
-  hideHippa,
-  formIsSigned,
-  formIsLocked
-}
